pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "delivery-verification-ml"
version = "0.1.0"
description = "HTTP microservice for driver verification: OCR + face similarity (+ forgery stub)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.6",
  "uvicorn[standard]>=0.32.1",
  "pydantic>=2.10.3",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.17",
  "numpy>=1.26.4",
  "opencv-python-headless>=4.10.0.84",
  "Pillow>=10.4.0",
  "rapidocr-onnxruntime>=1.4.4",
  "insightface>=0.7.3",
  "onnxruntime>=1.20.1",
]

[project.optional-dependencies]
dev = ["httpx>=0.28.0", "pytest>=8.3.0", "ruff>=0.8.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
